--- config/doctrine/Rappel.orm.yml
@@ -1,33 +1,0 @@
-App\Entity\Rappel:
-    type: entity
-    table: 'rappel'
-    repositoryClass: App\Repository\RappelRepository
-    id:
-        id:
-            type: integer
-            id: true
-            generator:
-                strategy: AUTO
-            options:
-                unsigned: true
-    fields:
-        date:
-            type: datetime
-        bEnvoye:
-            type: boolean
-        types:
-            type: array
-        archive:
-            type: boolean
-            options:
-                default: 0
-    manyToOne:
-        evenement:
-            targetEntity: App\Entity\Attributes\Evenement
-            inversedBy: rappels
-    oneToOne:
-        sms:
-            targetEntity: App\Entity\SMS
-            mappedBy: rappel
-            cascade: [ remove ]
-    lifecycleCallbacks: { }
