+++ src/Provider/EvenementProvider.php
@@ -5,7 +5,7 @@ namespace App\Provider;
 use App\Entity\Attributes\Client;
 use App\Entity\Attributes\DonneePersonnelle;
 use App\Entity\Attributes\Evenement;
-use App\Entity\Rappel;
+use App\Entity\Attributes\Rappel;
 use App\Event\DonneePersonnelleEvent;
 use App\Event\REEvent;
 use App\Form\Type\EvenementSimpleType;
