+++ src/Validator/Constraints/Rappel/EntityValidator.php
@@ -3,7 +3,7 @@
 namespace App\Validator\Constraints\Rappel;
 
 use App\Entity\Attributes\Beneficiaire;
-use App\Entity\Rappel;
+use App\Entity\Attributes\Rappel;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\Form\Exception\UnexpectedTypeException;
 use Symfony\Component\Validator\Constraint;
